Refactor dnb update admin tool to break out reusable functionality (#2525)
