+++ changelog/company/refactor-update-from-dnb-admin.internal.md
@@ -1,0 +1,3 @@
+The `update_from_dnb` admin tool was refactored to break out some useful utilities
+for use by other modules.  This will aid in the new "Link Company with D&B" tool
+which will follow shortly.
